Copy shareable URL and paste image
