+++ app/package.json
@@ -50,10 +50,11 @@
     "lucide-react": "^0.510.0",
     "next": "15.3.2",
     "next-themes": "^0.4.6",
+    "nuqs": "^2.6.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "react-chartjs-2": "^5.3.0",
     "react-day-picker": "^8.10.1",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.3.8",
     "react-hook-form": "^7.56.3",
     "react-resizable-panels": "^3.0.2",
