+++ app/src/app/dashboard/group/[id]/invoices/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { useParams } from "next/navigation";
+import { CompletedJobs } from "../completed-jobs";
+
+export default function InvoicesPage() {
+  const { id } = useParams();
+
+  return <CompletedJobs groupId={Number(id)} />;
+}
